refactor: Score 계산 역할 Hands로 변경
